Fix a bug when indexing a tx that burns Token Type 2 txs
We use the token ID 000...000 for invalid/unknown SLP txs, and this currently includes Token Type 2 txs.

This fixes a bug where a tx burns TT2 tokens breaks the indexer because it's trying to calculate token stats for the token with ID 000...000.

The fix is to skip burns that have a token ID of 000...000.

In the future, properly indexing Token Type 2 txs would be preferable.